start to refactor the ImagesListViewController on MVP
